+++ frontend/src/app/login/page.js
@@ -1,0 +1,39 @@
+'use client';
+import { useState } from 'react';
+
+export default function Login() {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [message, setMessage] = useState('');
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    setMessage('Login feature not implemented yet.');
+  };
+
+  return (
+    <main className="min-h-screen flex flex-col items-center justify-center p-8">
+      <h1 className="text-3xl font-bold mb-4">Log In</h1>
+      <form onSubmit={handleSubmit} className="space-y-4 w-full max-w-sm">
+        <input
+          type="email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          placeholder="Email"
+          className="w-full p-2 border rounded"
+          required
+        />
+        <input
+          type="password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          placeholder="Password"
+          className="w-full p-2 border rounded"
+          required
+        />
+        <button type="submit" className="w-full p-2 bg-indigo-600 text-white rounded">Log In</button>
+      </form>
+      {message && <p className="mt-4 text-sm text-gray-600">{message}</p>}
+    </main>
+  );
+}
